Type conversion fix

+++ zombie_game/Floor.cpp
@@ -1,6 +1,6 @@
 #include "Floor.h"
 
-Floor::Floor(AssetManager& assets, sf::Vector2f pos, sf::Vector2f size) : Static(assets)
+Floor::Floor(AssetManager& assets, sf::Vector2f pos, sf::Vector2i size) : Static(assets)
 {
 	this->sprite.setTexture(assets.floorTexture);
 	this->sprite.setTextureRect(sf::IntRect(0, 0, size.x, size.y));
